Prevent superfluous string comparison: Revert "Mira Cast Part 6/7"
At first, my strncmp-dar was firing, but it was a false alarm because that's a c++ string, not a char*.
Subsequently, I realized it's doing a string comparison against a magic in-line constant string every time updateAndReleaseLocked happens... and decided to put it out of its misery.

ChromeCast > miracast. They're cheap... and so are you if you miss miracast rotation.

This reverts commit cb21c420e17d24f5bc1a05144d27ac83087d9e01.

Change-Id: I57880a97ca73eb499795b047726797d1a3539c26